Document euclidean_distance function, add explanations
Since we now have two distance functions, I've clarified their
semantic differences a bit, and use a single shared example
page for them so they can be directly contrasted.

I've also changed the description of the arguments; the functions
don't care whether either argument is a column or a vector literal,
just that they have vectors of the same number of dimensions. (I say
"number of dimensions" rather than "length" because vectors have a
length in the sense of "distance from a vector with all components
0" and I want to avoid confusion.)