updated from the blog post, but needs review to ensure accuracy (#228)
This update was made because of the Azure announcement on the
FeatureBase website

https://www.featurebase.com/blog/multi-cloud-support-microsoft-azure

The content was copied from the following sections and may need
clarification and updating:

* Key differences of FeatureBase Azure vs FeatureBase AWS
* Meeting compliance... the second paragraph starting with "...The
standard configuration..."

Please check this PR for accuracy, **especially** for AWS for which I've
got no information

**NOTE** At time of writing (2024-01-16), Azure is not available in
FeatureBase Cloud. I expect this will change in the near future.